Refactor all testing classes/instantiations into testing package.
