+++ src/main/scala/com/itszuvalex/itszulib/testing/PortalTileTest.scala
@@ -1,4 +1,4 @@
-package com.itszuvalex.itszulib.render
+package com.itszuvalex.itszulib.testing
 
 import net.minecraft.tileentity.TileEntity
